docs: Update Request loaders guide (#1376)
- Adaptive crawler - prefer combining with Parsel.
- Highlight some important parts of the code samples.
- Prefer Impit over Httpx.
- Expose `FingerprintGenerator` in the public API docs.
- Add a note to the request loaders guide to highlight the usage with
crawlers.
