+++ docs/guides/code_examples/request_loaders/rl_basic_example.py
@@ -18,6 +18,7 @@ async def main() -> None:
     # Fetch and process requests from the queue.
     while request := await request_list.fetch_next_request():
         # Do something with it...
+        print(f'Processing {request.url}')
 
         # And mark it as handled.
         await request_list.mark_request_as_handled(request)
