+++ docs/guides/code_examples/request_loaders/rl_tandem_example.py
@@ -8,9 +8,11 @@ async def main() -> None:
     # Create a static request list.
     request_list = RequestList(['https://crawlee.dev', 'https://apify.com'])
 
+    # highlight-start
     # Convert the request list to a request manager using the to_tandem method.
     # It is a tandem with the default request queue.
     request_manager = await request_list.to_tandem()
+    # highlight-end
 
     # Create a crawler and pass the request manager to it.
     crawler = ParselCrawler(
@@ -20,8 +22,19 @@ async def main() -> None:
 
     @crawler.router.default_handler
     async def handler(context: ParselCrawlingContext) -> None:
+        context.log.info(f'Processing {context.request.url}')
+
         # New links will be enqueued directly to the queue.
         await context.enqueue_links()
+
+        # Extract data using Parsel's XPath and CSS selectors.
+        data = {
+            'url': context.request.url,
+            'title': context.selector.xpath('//title/text()').get(),
+        }
+
+        # Push extracted data to the dataset.
+        await context.push_data(data)
 
     await crawler.run()
