+++ docs/guides/code_examples/request_loaders/sitemap_tandem_example.py
@@ -2,38 +2,51 @@ import asyncio
 import re
 
 from crawlee.crawlers import ParselCrawler, ParselCrawlingContext
-from crawlee.http_clients import HttpxHttpClient
+from crawlee.http_clients import ImpitHttpClient
 from crawlee.request_loaders import SitemapRequestLoader
 
 
 async def main() -> None:
-    # Create an HTTP client for fetching sitemaps
-    async with HttpxHttpClient() as http_client:
-        # Create a sitemap request loader with URL filtering
-        sitemap_loader = SitemapRequestLoader(
-            sitemap_urls=['https://crawlee.dev/sitemap.xml'],
-            http_client=http_client,
-            # Include only URLs that contain 'docs'
-            include=[re.compile(r'.*docs.*')],
-            max_buffer_size=500,  # Buffer up to 500 URLs in memory
-        )
+    # Create an HTTP client for fetching the sitemap.
+    http_client = ImpitHttpClient()
 
-        # Convert the sitemap loader to a request manager using the to_tandem method.
-        # It is a tandem with the default request queue.
-        request_manager = await sitemap_loader.to_tandem()
+    # Create a sitemap request loader with filtering rules.
+    sitemap_loader = SitemapRequestLoader(
+        sitemap_urls=['https://crawlee.dev/sitemap.xml'],
+        http_client=http_client,
+        include=[re.compile(r'.*docs.*')],  # Only include URLs containing 'docs'.
+        max_buffer_size=500,  # Keep up to 500 URLs in memory before processing.
+    )
 
-        # Create a crawler and pass the request manager to it.
-        crawler = ParselCrawler(
-            request_manager=request_manager,
-            max_requests_per_crawl=10,  # Limit the max requests per crawl.
-        )
+    # highlight-start
+    # Convert the sitemap loader into a request manager linked
+    # to the default request queue.
+    request_manager = await sitemap_loader.to_tandem()
+    # highlight-end
 
-        @crawler.router.default_handler
-        async def handler(context: ParselCrawlingContext) -> None:
-            # New links will be enqueued directly to the queue.
-            await context.enqueue_links()
+    # Create a crawler and pass the request manager to it.
+    crawler = ParselCrawler(
+        request_manager=request_manager,
+        max_requests_per_crawl=10,  # Limit the max requests per crawl.
+    )
 
-        await crawler.run()
+    @crawler.router.default_handler
+    async def handler(context: ParselCrawlingContext) -> None:
+        context.log.info(f'Processing {context.request.url}')
+
+        # New links will be enqueued directly to the queue.
+        await context.enqueue_links()
+
+        # Extract data using Parsel's XPath and CSS selectors.
+        data = {
+            'url': context.request.url,
+            'title': context.selector.xpath('//title/text()').get(),
+        }
+
+        # Push extracted data to the dataset.
+        await context.push_data(data)
+
+    await crawler.run()
 
 
 if __name__ == '__main__':
